+++ android/src/main/java/com/sumsg/metronome/Metronome.java
@@ -15,11 +15,11 @@ public class Metronome {
     private short[] mTookSilenceSoundArray;
     private AudioGenerator.Sample mTook;
     private int mBeatDivisionSampleCount;
-    private float volume= (float) 0.5;
+    private float volume = (float) 0.5;
     private final Context context;
     private BeatTimer beatTimer;
 
-    public Metronome(Context ctx,String mainFile) {
+    public Metronome(Context ctx, String mainFile) {
         context = ctx;
         setAudioFile(mainFile);
     }
@@ -27,11 +27,13 @@ public class Metronome {
     public void setAudioFile(String path) {
         mTook = AudioGenerator.loadSampleFromWav(path);
     }
-    public void enableTapCallback(EventChannel.EventSink _eventTapSink) {
-        beatTimer = new BeatTimer(_eventTapSink);
+
+    public void enableTickCallback(EventChannel.EventSink _eventTickSink) {
+        beatTimer = new BeatTimer(_eventTickSink);
     }
+
     public void calcSilence() {
-        //(beats per second * SAMPLE_RATE) - NumberOfSamples
+        // (beats per second * SAMPLE_RATE) - NumberOfSamples
         mBeatDivisionSampleCount = (int) (((60 / (float) mBpm) * SAMPLE_RATE));
         int silence = Math.max(mBeatDivisionSampleCount - mTook.getSampleCount(), 0);
         mTookSilenceSoundArray = new short[silence];
@@ -41,7 +43,7 @@ public class Metronome {
     }
 
     private void isInitialized() {
-        if (mTook == null ) {
+        if (mTook == null) {
             throw new IllegalStateException("Not initialized correctly");
         }
     }
@@ -59,68 +61,78 @@ public class Metronome {
                 audioGenerator.writeSound(sample, Math.min(sample.length, mBeatDivisionSampleCount));
                 audioGenerator.writeSound(mTookSilenceSoundArray);
                 synchronized (mLock) {
-                    if (!play) return;
+                    if (!play)
+                        return;
                 }
             } while (true);
         }).start();
         //
-        if (beatTimer!=null){
+        if (beatTimer != null) {
             beatTimer.startBeatTimer(bpm);
         }
     }
+
     public void pause() {
-        if (audioGenerator.getAudioTrack()!=null){
+        if (audioGenerator.getAudioTrack() != null) {
             play = false;
             audioGenerator.getAudioTrack().pause();
         }
-        if (beatTimer!=null) {
+        if (beatTimer != null) {
             beatTimer.stopBeatTimer();
         }
     }
+
     public void stop() {
-        if (audioGenerator.getAudioTrack()!=null){
+        if (audioGenerator.getAudioTrack() != null) {
             play = false;
             audioGenerator.getAudioTrack().stop();
         }
-        if (beatTimer!=null) {
+        if (beatTimer != null) {
             beatTimer.stopBeatTimer();
         }
     }
+
     public int getVolume() {
         return (int) (volume * 100);
     }
+
     public void setVolume(float val) {
-        if (audioGenerator.getAudioTrack()!=null){
+        if (audioGenerator.getAudioTrack() != null) {
             volume = val;
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP ) {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
                 try {
                     audioGenerator.getAudioTrack().setVolume(volume);
-                }catch(Exception e) {
+                } catch (Exception e) {
                     Log.e("setVolume", String.valueOf(e));
                 }
             }
         }
     }
+
     public void setBPM(int bpm) {
         mBpm = bpm;
         calcSilence();
-        if (play){
-            if (beatTimer!=null) {
+        if (play) {
+            if (beatTimer != null) {
                 beatTimer.startBeatTimer(bpm);
             }
         }
     }
+
     public double getBPM() {
         return mBpm;
     }
+
     public boolean isPlaying() {
         return play;
     }
+
     public void destroy() {
-        if (beatTimer!=null) {
+        if (beatTimer != null) {
             beatTimer.stopBeatTimer();
         }
-        if (!play) return;
+        if (!play)
+            return;
         stop();
         synchronized (mLock) {
             mBeatDivisionSampleCount = 0;